Moves testhelpers Util to equalsverifier-testhelpers
