+++ equalsverifier-core/src/test/java/nl/jqno/equalsverifier/internal/util/ListBuildersTest.java
@@ -1,6 +1,6 @@
 package nl.jqno.equalsverifier.internal.util;
 
-import static nl.jqno.equalsverifier.internal.testhelpers.Util.coverThePrivateConstructor;
+import static nl.jqno.equalsverifier_testhelpers.Util.coverThePrivateConstructor;
 
 import org.junit.jupiter.api.Test;
