+++ equalsverifier-testhelpers/src/main/java/nl/jqno/equalsverifier_testhelpers/Util.java
@@ -1,4 +1,4 @@
-package nl.jqno.equalsverifier.internal.testhelpers;
+package nl.jqno.equalsverifier_testhelpers;
 
 import java.lang.reflect.Constructor;
